correccion excepciones, agregado readme, movida modelos de errores

+++ src/main/java/com/victoriap/desafioBackend/model/CrearProducto.java
@@ -18,5 +18,5 @@ public class CrearProducto {
     private String descripcion;
 
     @NotBlank
-    private float precio;
+    private double precio;
 }
